Get rid of ::serde use statement
