--- application/comit_node/src/settings/serde/mod.rs
@@ -1,3 +1,0 @@
-pub mod duration;
-pub mod log_level;
-pub mod url;
